pipeline/news-batch/requirements.airflow.txt
apache-airflow-providers-postgres
apache-airflow-providers-apache-spark
apache-airflow-providers-elasticsearch

pyarrow
python-dotenv
loguru

matplotlib
wordcloud
Pillow

